+++ packages/js-utils/src/Http.ts
@@ -1,0 +1,8 @@
+import { getKy, getShortUrl } from './Got/getKy'
+
+class HttpClass {
+  getClient = getKy
+  getShortUrl = getShortUrl
+}
+
+export const Http = new HttpClass()