+++ packages/node-utils/src/index.ts
@@ -9,8 +9,6 @@ import { Security } from './Security'
 import { runScript, Scripts } from './Scripts'
 import { Env } from './Env'
 import type { BuildInfo } from './Env'
-import { getGot } from './getGot'
-import { Inspect } from './Inspect'
 import { ObjectIds } from './ObjectIds'
 import { Streams } from './Streams'
 import { Ips } from './Ips'
@@ -27,8 +25,6 @@ export {
   Security,
   runScript,
   Env,
-  getGot,
-  Inspect,
   Scripts,
   ObjectIds,
   Streams,