+++ packages/node-xlsx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # 0.5.20
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.37
+
 ## 0.0.4
 
 ### Patch Changes